+++ src/onegov.org/onegov/org/assets/js/reservationcalendar_custom.js
@@ -12,7 +12,9 @@ var setupReservationCalendar = function(calendar) {
         highlights_min: calendar.data('highlights-min'),
         highlights_max: calendar.data('highlights-max'),
         reservations: calendar.data('reservations'),
-        reservationform: calendar.data('reservationform')
+        reservationform: calendar.data('reservationform'),
+        resourcesUrl: calendar.data('resources-url'),
+        resourceActive: calendar.data('resource-active')
     });
 };